refactor: Shop모듈이 모두 shop-app 모듈로 의존성이 향하게 한다
