+++ auth/auth-repository/build.gradle
@@ -1,8 +1,7 @@
 dependencies {
-    implementation project(':core:entity')
     implementation project(':auth:auth-domain')
 
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 
-    testImplementation 'com.h2database:h2:2.1.214' // TODO : 버전업
+    testImplementation 'com.h2database:h2'
 }
