--- core/entity/build.gradle
@@ -1,3 +1,0 @@
-dependencies {
-    implementation 'javax.persistence:javax.persistence-api:2.2'
-}
