+++ core/id-gen/src/test/java/shop/woowasap/core/id/snowflake/SnowflakeIdGeneratorTest.java
@@ -54,8 +54,7 @@ class SnowflakeIdGeneratorTest {
             // given
             final int expectedGeneratedIdCount = 10000;
             final ExecutorService executorService = Executors.newFixedThreadPool(256);
-            final List<Callable<Long>> idGenerateThreads = idGenerateThreads(
-                expectedGeneratedIdCount);
+            final List<Callable<Long>> idGenerateThreads = idGenerateThreads(expectedGeneratedIdCount);
 
             // when & then
             waitAtMost(ofSeconds(10))
