+++ coverage-exclude.asap
@@ -1,5 +1,4 @@
 exclude shop/woowasap/Application
-exclude shop/woowasap/shop/domain/*
 exclude shop/woowasap/auth/domain/*
 exclude shop/woowasap/shop/controller/*
 exclude shop/woowasap/shop/repository/*
