+++ shop/shop-app/src/main/java/shop/woowasap/shop/app/api/request/RegisterProductRequest.java
@@ -1,0 +1,8 @@
+package shop.woowasap.shop.app.api.request;
+
+import java.time.LocalDateTime;
+
+public record RegisterProductRequest(String name, String description, String price, long quantity,
+                                     LocalDateTime startTime, LocalDateTime endTime) {
+
+}
