+++ shop/shop-app/src/main/java/shop/woowasap/shop/app/exception/InvalidProductNameException.java
@@ -1,0 +1,5 @@
+package shop.woowasap.shop.app.exception;
+
+public class InvalidProductNameException extends ProductException {
+
+}
