+++ shop/shop-app/src/main/java/shop/woowasap/shop/app/exception/InvalidProductQuantityException.java
@@ -1,0 +1,5 @@
+package shop.woowasap.shop.app.exception;
+
+public class InvalidProductQuantityException extends ProductException {
+
+}
