+++ shop/shop-app/src/main/java/shop/woowasap/shop/app/exception/ProductException.java
@@ -1,0 +1,9 @@
+package shop.woowasap.shop.app.exception;
+
+public abstract class ProductException extends RuntimeException {
+
+    @Override
+    public synchronized Throwable fillInStackTrace() {
+        return this;
+    }
+}
