+++ shop/shop-app/src/main/java/shop/woowasap/shop/app/exception/UpdateProductException.java
@@ -1,0 +1,13 @@
+package shop.woowasap.shop.app.exception;
+
+public class UpdateProductException extends RuntimeException {
+
+    public UpdateProductException(String message) {
+        super(message);
+    }
+
+    @Override
+    public synchronized Throwable fillInStackTrace() {
+        return this;
+    }
+}
