+++ shop/shop-controller/build.gradle
@@ -1,5 +1,5 @@
 dependencies {
-    implementation project(':shop:shop-service')
+    implementation project(':shop:shop-app')
 
     implementation 'org.springframework.boot:spring-boot-starter-web'
 }
