--- shop/shop-domain/src/main/java/shop/woowasap/shop/domain/product/exception/InvalidProductNameException.java
@@ -1,5 +1,0 @@
-package shop.woowasap.shop.domain.product.exception;
-
-public class InvalidProductNameException extends ProductException {
-
-}
