--- shop/shop-domain/src/main/java/shop/woowasap/shop/domain/product/exception/InvalidProductQuantityException.java
@@ -1,5 +1,0 @@
-package shop.woowasap.shop.domain.product.exception;
-
-public class InvalidProductQuantityException extends ProductException {
-
-}
