--- shop/shop-domain/src/main/java/shop/woowasap/shop/domain/product/exception/ProductException.java
@@ -1,9 +1,0 @@
-package shop.woowasap.shop.domain.product.exception;
-
-public abstract class ProductException extends RuntimeException {
-
-    @Override
-    public synchronized Throwable fillInStackTrace() {
-        return this;
-    }
-}
