+++ shop/shop-repository/src/main/java/shop/woowasap/shop/repository/MockProductRepository.java
@@ -4,8 +4,8 @@ import java.time.LocalDateTime;
 import java.time.ZoneOffset;
 import java.util.Optional;
 import org.springframework.stereotype.Repository;
-import shop.woowasap.shop.domain.product.Product;
-import shop.woowasap.shop.service.repository.ProductRepository;
+import shop.woowasap.shop.app.domain.product.Product;
+import shop.woowasap.shop.app.spi.ProductRepository;
 
 @Repository
 public class MockProductRepository implements ProductRepository {
