--- shop/shop-service/src/main/java/shop/woowasap/shop/service/dto/RegisterProductRequest.java
@@ -1,8 +1,0 @@
-package shop.woowasap.shop.service.dto;
-
-import java.time.LocalDateTime;
-
-public record RegisterProductRequest(String name, String description, String price, long quantity,
-                                     LocalDateTime startTime, LocalDateTime endTime) {
-
-}
