--- shop/shop-service/src/main/java/shop/woowasap/shop/service/exception/UpdateProductException.java
@@ -1,13 +1,0 @@
-package shop.woowasap.shop.service.exception;
-
-public class UpdateProductException extends RuntimeException {
-
-    public UpdateProductException(String message) {
-        super(message);
-    }
-
-    @Override
-    public synchronized Throwable fillInStackTrace() {
-        return this;
-    }
-}
