+++ shop/shop-service/src/test/java/shop/woowasap/shop/service/ProductServiceTest.java
@@ -19,11 +19,11 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import shop.woowasap.core.id.api.IdGenerator;
-import shop.woowasap.shop.domain.product.Product;
-import shop.woowasap.shop.service.dto.RegisterProductRequest;
-import shop.woowasap.shop.service.dto.UpdateProductRequest;
-import shop.woowasap.shop.service.exception.UpdateProductException;
-import shop.woowasap.shop.service.repository.ProductRepository;
+import shop.woowasap.shop.app.domain.product.Product;
+import shop.woowasap.shop.app.api.request.RegisterProductRequest;
+import shop.woowasap.shop.app.api.request.UpdateProductRequest;
+import shop.woowasap.shop.app.exception.UpdateProductException;
+import shop.woowasap.shop.app.spi.ProductRepository;
 import shop.woowasap.shop.service.support.fixture.DomainFixture;
 
 @ExtendWith(SpringExtension.class)
